2025-12-01 rest層とdomain層を繋ぐコード追加実装、修正

+++ loginapp/src/main/java/com/example/loginapp/domain/model/Product.java
@@ -1,5 +1,7 @@
 package com.example.loginapp.domain.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -19,5 +21,6 @@ public class Product {
     private String name;
 
     /** 商品価格 */
-    private double price;
+    // private double price;
+    private BigDecimal price;
 }
