+++ loginapp/src/main/java/com/example/loginapp/usecase/product/UpdateTwoProductsInputData.java
@@ -1,5 +1,7 @@
 package com.example.loginapp.usecase.product;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -14,11 +16,11 @@ public class UpdateTwoProductsInputData {
     private int id1;
 
     /** 1 件目の商品に設定する新しい価格。 */
-    private double price1;
+    private BigDecimal price1;
 
     /** 2 件目に更新を行う対象商品の ID。 */
     private int id2;
 
     /** 2 件目の商品に設定する新しい価格。 */
-    private double price2;
+    private BigDecimal price2;
 }
